ExamOnlineBack/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==5.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-crispy-forms==1.14.0
django-filter==22.1
django-formtools==2.4
django-guardian==2.4.0
django-import-export==3.0.2
django-reversion==5.0.4
djangorestframework==3.14.0
djangorestframework-jwt-4==1.12.0
et-xmlfile==1.1.0
future==0.18.3
httplib2==0.9.2
idna==3.4
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openpyxl==3.1.0
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tablib==3.3.0
uritemplate==4.1.1
urllib3==1.26.14
xadmin2==2.0.4
xlrd==2.0.1
xlwt==1.3.0
pyqt5==5.15.9
pyqtwebengine==5.15.6